run: prepare API modules for independent evolution
Split run and result API views into dedicated modules to keep
entity-specific code isolated and easier to evolve independently.
